requirements.txt
# Core dependencies
loguru==0.7.2
praw==7.7.1
PyYAML==6.0.2
kafka-python==2.0.2

# Data processing & Spark (optional: Spark runtime provided separately)
pyspark==3.5.1
delta-spark==3.2.0

# Pandas UDF & Array / DataFrame ops
numpy>=1.25.0
pandas>=2.1.0
pyarrow>=15.0.0

# Machine Learning / Deep Learning
torch>=2.2.0
transformers>=4.41.0

# Development & Testing
pytest==8.3.2
ruff==0.6.8

# Notebook
matplotlib==3.8.0